requirements.txt
aiohttp==3.9.1
asyncio==3.4.3
boto3==1.34.14
feedparser==6.0.10
psycopg2-binary==2.9.9
python-dotenv==1.0.0
greenlet==2.0.2
